+++ RELEASENOTES.md
@@ -3,6 +3,8 @@
 * Added a progress bar with a configurable length.
 * Official internal code clean-up and splitting.
 * Added One Drive directory detection.
+* Added configurable page count.
+* Moved release notes and credits to their own page.
 
 # 0.1.1.0
 * Considerably quicker extracting speeds.